Refactor into package per parser
