+++ benchmark/src/main/java/com/novoda/jackson/JacksonSmallXmlBenchmark.java
@@ -1,4 +1,4 @@
-package com.novoda;
+package com.novoda.jackson;
 
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
